day16-fastapi-weather/requirements.txt
fastapi==0.95.2
uvicorn==0.23.0
requests==2.31.0
pytest==7.4.0
httpx==0.24.1
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.23.0
requests==2.31.0
pytest==7.4.0
httpx==0.24.1
python-dotenv==1.0.0
